[FIXED] Bug release by the author.
The product Duet Jekyll Theme has just been updated.

The seller has added the following message for this update:

This update fixes a small bug related to the masonry grid layout on slower connections or where there are a lot of posts.

To update your project, simply replace the js/duet-min.js and js/duet.js files with the new ones in this update.

Thanks!